App/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
requests>=2.28.0
sentence-transformers>=2.2.2
openai>=1.0.0
networkx>=3.1
scikit-learn>=1.3.0
pandas>=2.0.0
aiohttp>=3.8.0
alembic>=1.12.0
torch>=2.0.0
numpy>=1.24.0
logging>=0.5.1.2